Moved UserCreated to folder/namespace for external events
