+++ Source/Admin/Web/TestDataGeneratorController.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Read;
 using MongoDB.Driver;
 using Newtonsoft.Json;
+using Events.External;
 
 namespace Web
 {
